Move screen into own package
