+++ varp-mod/varp-platform-fabric/src/client/kotlin/net/voxelpi/varp/mod/fabric/client/gui/screen/FabricVarpExplorerScreen.kt
@@ -1,4 +1,4 @@
-package net.voxelpi.varp.mod.fabric.client.gui
+package net.voxelpi.varp.mod.fabric.client.gui.screen
 
 import io.wispforest.owo.ui.base.BaseOwoScreen
 import io.wispforest.owo.ui.container.Containers
